refactor(features): move features in components/appli
