+++ packages/web-app/src/components/appli/AuthChecker/index.jsx
@@ -6,9 +6,9 @@ import { useIntl } from 'react-intl';
 import { useDispatch } from 'react-redux';
 import styled from 'styled-components';
 
-import { displayLoginDialog } from '../../actions/Login';
-import ErrorMessage from '../../components/common/StatusMessage/ErrorMessage';
-import { usePermissions } from '../../hooks';
+import { displayLoginDialog } from '../../../actions/Login';
+import ErrorMessage from '../../common/StatusMessage/ErrorMessage';
+import { usePermissions } from '../../../hooks';
 
 // ====================
