+++ packages/web-app/src/components/appli/DocumentLanguageSelect.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 
-import LanguageSelect from '../components/appli/Document/DocumentForm/formElements/LanguageSelect';
+import LanguageSelect from './Document/DocumentForm/formElements/LanguageSelect';
 
 const DocumentLanguageSelect = ({
   contextValueName,
