+++ packages/web-app/src/components/appli/ForgotPassword.jsx
@@ -4,10 +4,10 @@ import { useIntl } from 'react-intl';
 import { isEmpty, match } from 'ramda';
 
 import { useHistory } from 'react-router-dom';
-import { emailRegexp } from '../conf/Config';
-import { postForgotPassword } from '../actions/ForgotPassword';
-import { useBoolean, useNotification, usePermissions } from '../hooks';
-import ForgotPasswordPage from '../pages/ForgotPassword';
+import { emailRegexp } from '../../conf/Config';
+import { postForgotPassword } from '../../actions/ForgotPassword';
+import { useBoolean, useNotification, usePermissions } from '../../hooks';
+import ForgotPasswordPage from '../../pages/ForgotPassword';
 
 const ForgotPassword = () => {
   const dispatch = useDispatch();
