+++ packages/web-app/src/components/appli/Form/MassifAutoComplete.jsx
@@ -7,11 +7,11 @@ import { InputAdornment } from '@material-ui/core';
 import {
   fetchQuicksearchResult,
   resetQuicksearch
-} from '../../actions/Quicksearch';
+} from '../../../actions/Quicksearch';
 
-import { entityOptionForSelector } from '../../helpers/Entity';
+import { entityOptionForSelector } from '../../../helpers/Entity';
 
-import SearchBar from '../../components/appli/Document/DocumentForm/formElements/SearchBar';
+import SearchBar from '../Document/DocumentForm/formElements/SearchBar';
 import FormAutoComplete from './FormAutoComplete';
 
 // ===================================
