+++ packages/web-app/src/components/appli/SignUp.jsx
@@ -4,10 +4,10 @@ import { useIntl } from 'react-intl';
 import { isEmpty, match } from 'ramda';
 
 import { useHistory } from 'react-router-dom';
-import { emailRegexp, PASSWORD_MIN_LENGTH } from '../conf/Config';
-import { postSignUp } from '../actions/SignUp';
-import { useNotification, usePermissions } from '../hooks';
-import SignUpForm from '../pages/SignUpForm';
+import { emailRegexp, PASSWORD_MIN_LENGTH } from '../../conf/Config';
+import { postSignUp } from '../../actions/SignUp';
+import { useNotification, usePermissions } from '../../hooks';
+import SignUpForm from '../../pages/SignUpForm';
 
 const SignUp = () => {
   const dispatch = useDispatch();
